Updated the code according to the changes requested by @aldeed
